requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0  
pytest==7.4.2
pytest-flask==1.2.0
bandit==1.7.5